Extract empty channel to separate module
