+++ src/components/channel/index.tsx
@@ -1,11 +1,10 @@
 import { observer } from "mobx-react";
-import React, { ChangeEvent, useRef } from "react";
+import React, { useRef } from "react";
 import { useEffect } from "react";
 import { useState } from "react";
-import { FileInputId } from "../../constants";
 import { Channel } from "../../models/channels";
 import Mixer from "../../models/mixer";
-import { ChannelDto } from "../../types";
+import EmptyChannel from "../emptyChannel";
 import editModeStore from "../mixerActions/store";
 import ChannelFader from "./addOns/channelFader";
 import ChannelName from "./addOns/channelName";
@@ -21,70 +20,6 @@ const LoadingSpinner: React.FC = () => {
   );
 };
 
-interface EmptyChannelProps {
-  index: number;
-  selectedIndex?: number;
-  onClick: Function;
-  onFileSelect: (index: number, dto: ChannelDto) => void;
-}
-
-const EmptyChannel: React.FC<EmptyChannelProps> = (props) => {
-  const setShowingIndex = props.onClick;
-  const [show, setShow] = useState(false);
-  const fileInputId = `${FileInputId}-${props.index}`;
-
-  useEffect(() => {
-    setShow(props.index === props.selectedIndex);
-  }, [props]);
-
-  const handleFileSelect = (e: ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-
-    if (files) {
-      props.onFileSelect(props.index, {
-        title: files[0]?.name,
-        src: files[0],
-      });
-
-      e.target.value = "";
-    } else {
-      return;
-    }
-  };
-
-  return (
-    <S.Channel>
-      <input type="file" accept=".mp3,.wav" id={fileInputId} onChange={handleFileSelect} />
-      <S.EmptyChannel
-        onClick={(e) => {
-          e.stopPropagation();
-          setShowingIndex(props.index);
-        }}
-      >
-        {show ? (
-          <S.AddChannelButtonsContainer
-            onClick={(e) => {
-              e.stopPropagation();
-              setShowingIndex(undefined);
-            }}
-          >
-            <S.AddFileButton htmlFor={fileInputId} />
-            <S.ButtonsDivider />
-            <S.LinkFileButton htmlFor={fileInputId} />
-          </S.AddChannelButtonsContainer>
-        ) : null}
-
-        <S.EmptyChannelInner>
-          <S.EmptyChannelTextContainer>
-            <S.EmptyChannelText>CHANNEL</S.EmptyChannelText>
-            <S.EmptyChannelNumberText>{props.index + 1}</S.EmptyChannelNumberText>
-          </S.EmptyChannelTextContainer>
-        </S.EmptyChannelInner>
-      </S.EmptyChannel>
-    </S.Channel>
-  );
-};
-
 interface ChannelComponentProps {
   channel: Channel;
   pressedKey?: string;
@@ -96,21 +31,19 @@ const ChannelComponent: React.FC<ChannelComponentProps> = observer(
     const store = editModeStore;
 
     return (
-      <S.Channel>
-        <S.ChannelInnerWrapper>
-          {channel.loaded ? null : <S.LoadingMask />}
-          <S.ChannelUserInfoSection>
-            {channel.loaded ? <S.ChannelUserProfileImg /> : <LoadingSpinner />}
-            {channel.loaded && store.isEditing ? (
-              <S.DeleteChannelButton onClick={() => onDeleteChannel(channel)} />
-            ) : null}
-          </S.ChannelUserInfoSection>
-          <MuteSoloComponent channel={channel} />
-          <Panner channel={channel} pressedKey={pressedKey} />
-          <ChannelFader channel={channel} pressedKey={pressedKey} />
-          <ChannelName channel={channel} />
-        </S.ChannelInnerWrapper>
-      </S.Channel>
+      <S.ChannelInnerWrapper>
+        {channel.loaded ? null : <S.LoadingMask />}
+        <S.ChannelUserInfoSection>
+          {channel.loaded ? <S.ChannelUserProfileImg /> : <LoadingSpinner />}
+          {channel.loaded && store.isEditing ? (
+            <S.DeleteChannelButton onClick={() => onDeleteChannel(channel)} />
+          ) : null}
+        </S.ChannelUserInfoSection>
+        <MuteSoloComponent channel={channel} />
+        <Panner channel={channel} pressedKey={pressedKey} />
+        <ChannelFader channel={channel} pressedKey={pressedKey} />
+        <ChannelName channel={channel} />
+      </S.ChannelInnerWrapper>
     );
   }
 );
@@ -152,22 +85,26 @@ const ChannelsContainer: React.FC<ChannelsContainerProps> = observer((props) => 
     return channels.map((channel, index) => {
       if (channel) {
         return (
-          <ChannelComponent
-            key={index}
-            channel={channel}
-            pressedKey={props.pressedKey}
-            onDeleteChannel={handleChannelDelete}
-          />
+          <S.Channel>
+            <ChannelComponent
+              key={index}
+              channel={channel}
+              pressedKey={props.pressedKey}
+              onDeleteChannel={handleChannelDelete}
+            />
+          </S.Channel>
         );
       } else {
         return (
-          <EmptyChannel
-            key={index}
-            index={index}
-            selectedIndex={selectedIndex}
-            onClick={setSelectedIndex}
-            onFileSelect={mixer.addChannel.bind(mixer)}
-          />
+          <S.Channel>
+            <EmptyChannel
+              key={index}
+              index={index}
+              selectedIndex={selectedIndex}
+              onClick={setSelectedIndex}
+              onFileSelect={mixer.addChannel.bind(mixer)}
+            />
+          </S.Channel>
         );
       }
     });
