+++ src/components/emptyChannel/index.tsx
@@ -1,0 +1,68 @@
+import { ChangeEvent, useEffect, useState } from "react";
+import { FileInputId } from "../../constants";
+import { ChannelDto } from "../../types";
+import * as S from "./styles";
+
+interface EmptyChannelProps {
+  index: number;
+  selectedIndex?: number;
+  onClick: Function;
+  onFileSelect: (index: number, dto: ChannelDto) => void;
+}
+
+const EmptyChannel: React.FC<EmptyChannelProps> = (props) => {
+  const setShowingIndex = props.onClick;
+  const [show, setShow] = useState(false);
+  const fileInputId = `${FileInputId}-${props.index}`;
+
+  useEffect(() => {
+    setShow(props.index === props.selectedIndex);
+  }, [props]);
+
+  const handleFileSelect = (e: ChangeEvent<HTMLInputElement>) => {
+    const files = e.target.files;
+
+    if (files) {
+      props.onFileSelect(props.index, {
+        title: files[0]?.name,
+        src: files[0],
+      });
+
+      e.target.value = "";
+    } else {
+      return;
+    }
+  };
+
+  return (
+    <S.EmptyChannel
+      onClick={(e) => {
+        e.stopPropagation();
+        setShowingIndex(props.index);
+      }}
+    >
+      {show ? (
+        <S.AddChannelButtonsContainer
+          onClick={(e) => {
+            e.stopPropagation();
+            setShowingIndex(undefined);
+          }}
+        >
+          <S.AddFileButton htmlFor={fileInputId} />
+          <input type="file" accept=".mp3,.wav" id={fileInputId} onChange={handleFileSelect} />
+          <S.ButtonsDivider />
+          <S.LinkFileButton />
+        </S.AddChannelButtonsContainer>
+      ) : null}
+
+      <S.EmptyChannelInner>
+        <S.EmptyChannelTextContainer>
+          <S.EmptyChannelText>CHANNEL</S.EmptyChannelText>
+          <S.EmptyChannelNumberText>{props.index + 1}</S.EmptyChannelNumberText>
+        </S.EmptyChannelTextContainer>
+      </S.EmptyChannelInner>
+    </S.EmptyChannel>
+  );
+};
+
+export default EmptyChannel;
